+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal/show.erb
@@ -1,8 +1,8 @@
 <h4 class="heading4 vote-title"><%= t("decidim.proposals.proposal_m.modal.title") %></h4>
 <div class="flex--sbc">
-  <%= vote_block_for(proposal, 1, "green") %>
+  <%= vote_block_for(proposal, 3, "green") %>
   <%= vote_block_for(proposal, 2, "yellow") %>
-  <%= vote_block_for(proposal, 3, "red") %>
+  <%= vote_block_for(proposal, 1, "red") %>
 </div>
 
 <%= action_authorized_link_to :vote,
