+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal_cell.rb
@@ -3,19 +3,11 @@
 module Decidim
   module DecidimAwesome
     module Voting
-      class ThreeFlagsProposalCell < Decidim::ViewModel
-        include Decidim::ComponentPathHelper
-        include Decidim::IconHelper
-        include Decidim::Proposals::Engine.routes.url_helpers
-
+      class ThreeFlagsProposalCell < ThreeFlagsBaseCell
         VOTE_WEIGHTS = [0, 1, 2, 3].freeze
 
         def show
           render :show
-        end
-
-        def proposal
-          model
         end
 
         def vote_block_for(proposal, weight, color)
@@ -38,10 +30,6 @@ module Decidim
           options[:from_proposals_list]
         end
 
-        def current_component
-          proposal.component
-        end
-
         def proposal_vote_path(weight)
           proposal_proposal_vote_path(proposal_id: proposal.id, from_proposals_list: from_proposals_list, weight: weight)
         end
@@ -52,16 +40,6 @@ module Decidim
 
         def voted_for_any?
           VOTE_WEIGHTS.any? { |opt| voted_for?(opt) }
-        end
-
-        # def opacity_class_for_abstain
-        #   voted_for_any? ? "semi-opaque" : "fully-opaque"
-        # end
-
-        private
-
-        def current_vote
-          @current_vote ||= Decidim::Proposals::ProposalVote.find_by(author: current_user, proposal: model)
         end
       end
     end
