+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal_modal/show.erb
@@ -1,15 +1,4 @@
-<!--<div class="reveal vote_proposal_modal voting-three-flag" id="<%#= modal_id %>" data-reveal role="dialog" aria-modal="true" aria-labelledby="<%#= modal_id %>-label">-->
-<!--  <div class="reveal__header">-->
-<!--    <h3 id="<%#= modal_id %>-label" class="reveal__title"><%#= translated_attribute(model.title) %></h3>-->
-<!--    <button class="close-button" data-close aria-label="Close" type="button">-->
-<!--      <span aria-hidden="true">&times;</span>-->
-<!--    </button>-->
-<!--  </div>-->
-<!--  <div id="proposal-<%#= model.id %>-vote-button">-->
-<%#= cell("decidim/decidim_awesome/voting/three_flags_proposal", model, proposal: model, from_proposals_list: true) %>
-<!--  </div>-->
-<!--</div>-->
-
+<!--the modal window will be used in another pr-->
 <div class="reveal vote_proposal_modal voting-three-flag" id="<%= modal_id %>" data-reveal role="dialog" aria-modal="true" aria-labelledby="<%#= modal_id %>-label">
   <div class="reveal__content">
     <p><%= vote_instructions %></p>
