+++ app/cells/decidim/decidim_awesome/voting/three_flags_proposal_modal_cell.rb
@@ -3,17 +3,11 @@
 module Decidim
   module DecidimAwesome
     module Voting
-      class ThreeFlagsProposalModalCell < Decidim::ViewModel
-        include Decidim::ComponentPathHelper
-        include Decidim::Proposals::Engine.routes.url_helpers
+      class ThreeFlagsProposalModalCell < ThreeFlagsBaseCell
         include Decidim::Proposals::Engine.routes.url_helpers
 
         def show
           render :show
-        end
-
-        def proposal
-          model
         end
 
         def modal_id
@@ -22,10 +16,6 @@ module Decidim
 
         def from_proposals_list
           options[:from_proposals_list]
-        end
-
-        def current_component
-          proposal.component
         end
 
         def vote_instructions
