+++ lib/decidim/decidim_awesome/engine.rb
@@ -103,7 +103,7 @@ module Decidim
             # voting.show_votes_count_view = "" # hide votes count if needed
             voting.proposal_m_cell_footer = "decidim/decidim_awesome/voting/three_flags/proposal_m_cell_footer"
             voting.weight_validator do |weight, _context|
-              weight.in? [1, 2, 3]
+              weight.in? [0, 1, 2, 3]
             end
           end
